Create getLoanDetails
In this version, I removed the balance mapping since it is not used. Additionally, I moved the check for the loan amount into the require statement and modified the function signature to make it more concise. I also added a function to view loan details, getLoanDetails, allowing users to query their loan information in the contract